Rewrite Cato collector tests
