+++ src/go/plugin/go.d/collector/cato_networks/topology_store_test.go
@@ -14,17 +14,41 @@ import (
 )
 
 func TestTopologyStorePublishesCurrentSnapshot(t *testing.T) {
-	var store topologyStore
+	tests := map[string]struct {
+		setup func(*topologyStore) *topology.Data
+		check func(*testing.T, *topologyStore, *topology.Data)
+	}{
+		"empty store has no topology": {
+			check: func(t *testing.T, store *topologyStore, _ *topology.Data) {
+				got, ok := store.CurrentTopology()
+				require.Nil(t, got)
+				require.False(t, ok)
+			},
+		},
+		"load returns published snapshot": {
+			setup: func(store *topologyStore) *topology.Data {
+				data := &topology.Data{Source: topologySource}
+				store.Publish(data)
+				return data
+			},
+			check: func(t *testing.T, store *topologyStore, data *topology.Data) {
+				got, ok := store.CurrentTopology()
+				require.True(t, ok)
+				require.Same(t, data, got)
+			},
+		},
+	}
 
-	_, ok := store.CurrentTopology()
-	require.False(t, ok)
-
-	data := &topology.Data{Source: topologySource}
-	store.Publish(data)
-
-	got, ok := store.CurrentTopology()
-	require.True(t, ok)
-	require.Same(t, data, got)
+	for name, tc := range tests {
+		t.Run(name, func(t *testing.T) {
+			var store topologyStore
+			var data *topology.Data
+			if tc.setup != nil {
+				data = tc.setup(&store)
+			}
+			tc.check(t, &store, data)
+		})
+	}
 }
 
 func TestTopologyStoreConcurrentPublishLoad(t *testing.T) {
@@ -61,10 +85,33 @@ func TestTopologyStoreConcurrentPublishLoad(t *testing.T) {
 }
 
 func TestFuncDepsAdapterWithoutTopologyStore(t *testing.T) {
-	var deps funcDepsAdapter
+	tests := map[string]struct {
+		deps  funcDepsAdapter
+		check func(*testing.T, *topology.Data, bool)
+	}{
+		"nil store reports unavailable": {
+			check: func(t *testing.T, data *topology.Data, ok bool) {
+				require.Nil(t, data)
+				require.False(t, ok)
+			},
+		},
+		"store delegates current topology": {
+			deps: func() funcDepsAdapter {
+				var store topologyStore
+				store.Publish(&topology.Data{Source: topologySource})
+				return funcDepsAdapter{store: &store}
+			}(),
+			check: func(t *testing.T, data *topology.Data, ok bool) {
+				require.True(t, ok)
+				require.Equal(t, topologySource, data.Source)
+			},
+		},
+	}
 
-	data, ok := deps.CurrentTopology()
-
-	require.Nil(t, data)
-	require.False(t, ok)
+	for name, tc := range tests {
+		t.Run(name, func(t *testing.T) {
+			data, ok := tc.deps.CurrentTopology()
+			tc.check(t, data, ok)
+		})
+	}
 }
